Created parsable syllable-word list, added fkgl column, count syllables, and others

+++ code/preprocess_pipeline.py
@@ -6,6 +6,9 @@ import pandas as pd
 import spacy
 import string
 from nltk.corpus import stopwords
+import sys
+from os.path import exists
+
 
 import re
 stop = stopwords.words('english')
@@ -37,30 +40,45 @@ def required_preprocess(essays):
 
 # Pipeline for feature extraction
 def feature_extract(essays):
-    essays = process_column(essays, 'TEXT', 'num_of_char', num_of_char) # Always missing 2568
-    essays = process_column(essays, 'TEXT', 'num_any_words', num_any_word) # Correct 658
-    essays = process_column(essays, 'TEXT','num_long_words', num_long_words) # Correct 112
-    # essays = process_column(essays, 'TEXT','num_short_words', num_short_words) 
 
-    # essays = process_column(essays, 'TEXT','num_sentence', num_sentences)
-    # essays = process_column(essays, 'TEXT', "num_diff_word_stop", num_diff_word_stop)
-    # essays = process_column(essays, 'TEXT', "num_diff_word_nstop", num_diff_word_nstop)
-    # essays = process_column(essays, 'TEXT', 'avg_sentence_length',avg_sentence_length)
+    
+    essays = process_column(essays, 'TEXT', 'num_of_char', num_of_char) # Always missing 2 2568
+    essays = process_column(essays, 'TEXT', 'num_any_words', num_any_word) # Correct 658, always 1 off
+    essays = process_column(essays, 'TEXT','num_long_words', num_long_words) # Correct 109, 3 off, 112 after diff check
+    essays = process_column(essays, 'TEXT','num_short_words', num_short_words) # Counterd 548, 2 off
+
+    essays = process_column(essays, 'TEXT','num_sentences', num_sentences) # Correct 38
+    essays = process_column(essays, 'TEXT', "num_diff_word_stop", num_diff_word_stop) # Correct 272
+    essays = process_column(essays, 'TEXT', "num_diff_word_nstop", num_diff_word_nstop)
+    essays = process_column(essays, 'TEXT', 'avg_sentence_length',avg_sentence_length)
+    essays = process_column(essays, 'TEXT', 'avg_word_length',avg_word_length)
+
+    essays = process_column(essays, 'TEXT', 'num_syllables',num_syllables)
+
+    essays = process_column(essays, 'TEXT', 'most_freq_word_length',most_freq_word_length) 
+    essays = process_column(essays, 'TEXT', 'most_freq_sentence_length',most_freq_sentence_length)
+
+    # Readability
+    essays = process_column(essays, 'TEXT', 'flesch_kincaid_grade_level',flesch_kincaid_grade_level)
+
 
     return essays
 
 def main():
-    essays = pd.read_csv('essays.csv', encoding='cp1252')
-    processed_essays = required_preprocess(essays)
-
-    processed_essays.to_csv('pre_processed.csv', sep=',', encoding='utf-8', index = False)
-
-    test_line = processed_essays['TEXT'][0]
+    processed_essays = None
+    if (not exists('pre_processed.csv')):
+        essays = pd.read_csv('essays.csv', encoding='cp1252')
+        processed_essays = required_preprocess(essays)
+        processed_essays.to_csv('pre_processed.csv', sep=',', encoding='utf-8', index = False)
+    else:
+        processed_essays = pd.read_csv('pre_processed.csv', encoding='utf-8')
     
+    test_line = processed_essays['TEXT'][5]
+
+    print(flesch_kincaid_grade_level(test_line))
     final_processed = feature_extract(processed_essays)
 
-    # print((num_long_words(test_line)))
-    print(final_processed.head(1))
+    print(final_processed.head(10))
 
     # final_processed.to_csv('final.csv', sep=',', encoding='utf-8', index = False)
